Separate low-level logic from the ViewModel
